pyproject.toml
[project]
name = "transport_management"
version = "0.1.0"
description = "swida assignment app"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.8",
]
